Save group config before creating new profile
This will prevent the profile being overwritten later if we fail in the
process of adding it. It happens because `group.LastId + 1` will be the
new profile ID, so if we use it, create a new json file, but then fail
to increament it and save it to group config, the next profile will be
overwriting the last one. Therefore we save the group config first